Gestion de roles y permisos con y sin @can

+++ app/Http/Controllers/VehiculoController.php
@@ -4,6 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Models\Vehiculo;
 use Illuminate\Http\Request;
+use App\Models\Rol;
+use App\Models\Permiso;
+use Illuminate\Support\Facades\Gate;
 
 class VehiculoController extends Controller
 {
@@ -12,8 +15,14 @@ class VehiculoController extends Controller
      */
     public function index()
     {
-        $vehiculos = Vehiculo::all();
-        return view('vehiculos.index', compact('vehiculos'));
+        // Verificar si el usuario autenticado tiene permiso para leer vehículos
+        if (Gate::allows('read-vehicles')) {
+            $vehiculos = Vehiculo::all();
+            return view('vehiculos.index', compact('vehiculos'));
+        } else {
+            // Mostrar un mensaje de error
+            return response()->json(['error' => 'No tienes permiso para acceder a esta ruta'], 403);
+        }
     }
 
     /**
@@ -21,7 +30,15 @@ class VehiculoController extends Controller
      */
     public function create()
     {
-        return view('vehiculos.create');
+        // Verificar si el usuario autenticado tiene permiso para crear vehículos
+        if (Gate::allows('create-vehicles')) {
+            return view('vehiculos.create');
+        } else {
+            // Mostrar un mensaje de error
+            //return response()->json(['error' => 'No tienes permiso para acceder a esta ruta'], 403);
+            return redirect()->route('vehiculos.index')->with('error', 'No puede crear vehículos.');
+
+        }
     }
 
     /**
@@ -47,7 +64,14 @@ class VehiculoController extends Controller
      */
     public function show(Vehiculo $vehiculo)
     {
-        return view('vehiculos.show', compact('vehiculo'));
+        // Verificar si el usuario autenticado tiene permiso para leer vehículos
+        if (Gate::allows('read-vehicles')) {
+            return view('vehiculos.show', compact('vehiculo'));
+        } else {
+            // Mostrar un mensaje de error
+            return response()->json(['error' => 'No tienes permiso para acceder a esta ruta'], 403);
+
+        }
     }
 
     /**
@@ -55,7 +79,14 @@ class VehiculoController extends Controller
      */
     public function edit(Vehiculo $vehiculo)
     {
-        return view('vehiculos.edit', compact('vehiculo'));
+        // Verificar si el usuario autenticado tiene permiso para actualizar vehículos con Gate
+        if (Gate::allows('update-vehicles')) {
+            return view('vehiculos.edit', compact('vehiculo'));
+        } else {
+            // Mostrar un mensaje de error
+            return response()->json(['error' => 'No tienes permiso para acceder a esta ruta'], 403);
+
+        }
     }
 
     /**
@@ -63,25 +94,38 @@ class VehiculoController extends Controller
      */
     public function update(Request $request, Vehiculo $vehiculo)
     {
-        $request->validate([
-            'placa' => 'required|unique:vehiculos,placa,' . $vehiculo->id,
-            'marca' => 'required',
-            'modelo' => 'required',
-            'año' => 'required|digits:4|integer',
-            'estado' => 'required|in:activo,inactivo',
-        ]);
-
-        $vehiculo->update($request->all());
-
-        return redirect()->route('vehiculos.index')->with('success', 'El vehículo ha sido actualizado exitosamente.');
+        // Verificar si el usuario autenticado tiene permiso para actualizar vehículos con Gate
+        if (Gate::allows('update-vehicles', $vehiculo)) {
+            $request->validate([
+                'placa' => 'required|unique:vehiculos,placa,' . $vehiculo->id,
+                'marca' => 'required',
+                'modelo' => 'required',
+                'año' => 'required|digits:4|integer',
+                'estado' => 'required|in:activo,inactivo',
+            ]);
+    
+            $vehiculo->update($request->all());
+    
+            return redirect()->route('vehiculos.index')->with('success', 'El vehículo ha sido actualizado exitosamente.');
+        } else {
+            // Mostrar un mensaje de error
+            return response()->json(['error' => 'No tienes permiso para acceder a esta ruta'], 403);
+        }
     }
 
     /**
      * Remove the specified resource from storage.
      */
+
     public function destroy(Vehiculo $vehiculo)
     {
-        $vehiculo->delete();
-        return redirect()->route('vehiculos.index')->with('success', 'Vehículo eliminado exitosamente.');
+        // Verificar si el usuario autenticado tiene permiso para eliminar vehículos
+        if (Gate::allows('delete-vehicles', $vehiculo)) {
+            $vehiculo->delete();
+            return redirect()->route('vehiculos.index')->with('success', 'Vehículo eliminado exitosamente.');
+        } else {
+            // Mostrar un mensaje de error
+            return redirect()->route('vehiculos.index')->with('error', 'No tiene permiso para eliminar registros.');
+        }
     }
 }
